HashMap

+++ src/discover/algorithm/elementary/array/SingleNumber.java
@@ -3,17 +3,18 @@ package discover.algorithm.elementary.array;
 import java.util.ArrayList;
 import java.util.Arrays;
 
+
+/**
+ * 给定一个非空整数数组，除了某个元素只出现一次以外，其余每个元素均出现两次
+ * 找出那个只出现了一次的元素
+ * 使用O(n)时间复杂度和O(1)空间复杂度
+ */
 public class SingleNumber {
 
     /**
-     * 给定一个非空整数数组，除了某个元素只出现一次以外，其余每个元素均出现两次
-     * 找出那个只出现了一次的元素
-     * 使用O(n)时间复杂度和O(1)空间复杂度
+     * 解法1 6ms 超过41.02%
      */
     public int singleNumber(int[] nums) {
-        /**
-         * 解法1 6ms 超过41.02%
-         */
         if (nums.length == 1) return nums[0];
         Arrays.sort(nums);
         for (int i = 0; i < nums.length; i++) {
@@ -25,11 +26,11 @@ public class SingleNumber {
         return -1;
     }
 
+    /**
+     * 解法2 1ms
+     * 使用位异或运算
+     */
     public int singleNumber2(int[] nums) {
-        /**
-         * 解法2 1ms
-         * 使用位异或运算
-         */
         for (int num : nums) {
             nums[0] ^= num;
         }
@@ -37,10 +38,10 @@ public class SingleNumber {
         return nums[0];
     }
 
+    /**
+     * 解法3 0ms
+     */
     public int singleNumber3(int[] nums) {
-        /**
-         * 解法3 0ms
-         */
         for (int i = 0; i < nums.length; i += 2) {
             nums[0] ^= nums[i] ^ nums[i + 1];
         }
@@ -48,20 +49,19 @@ public class SingleNumber {
         return nums[0];
     }
 
+    /**
+     * 三个月之前写的代码 141ms 10.33%
+     */
     public int singleNumber4(int[] nums) {
-        /**
-         * 三个月之前写的代码 141ms 10.33%
-         */
-        if (nums.length == 1)
-            return nums[0];
+        if (nums.length == 1) return nums[0];
+
         ArrayList<Integer> temp = new ArrayList<>();
+
         for (int num : nums) {
-            if (temp.contains(num))
-                temp.remove((Integer) num);
-            else
-                temp.add(num);
+            if (temp.contains(num)) temp.remove((Integer) num);
+            else temp.add(num);
         }
 
-        return (int) temp.get(0);
+        return temp.get(0);
     }
 }
